Ml-ops(project)/GunDetect/requirements.txt
numpy==1.24.3
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.0
python-multipart==0.0.6
python-dotenv==1.0.0